Change interaction target finding to use raycast
Modified the `FindInteractionTarget` method in the `Interactor` class to use a raycast from the center of the screen to find the interactable object the player is looking at within range, instead of calculating the distance to all interactable objects. Updated the summary and return documentation comments to reflect this change.